requirements.txt
asgiref==3.8.1
Django==5.1.3
factory_boy==3.3.1
Faker==30.8.2
mysqlclient==2.2.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
